XW-30 unify naming conventions after XW-474 merge
