+++ server/facets/operations/prepareArticleData.ts
@@ -51,7 +51,7 @@ function prepareArticleData(request: Hapi.Request, data: ArticlePageData): any {
 	}
 
 	result.displayTitle = title;
-	result.isMainPage = (title === wikiVariables.mainPageTitle.replace(/_/g, ' '));
+	result.isMainPage = articleData.isMainPage;
 	result.canonicalUrl = wikiVariables.basePath + wikiVariables.articlePath + title.replace(/ /g, '_');
 	result.themeColor = Utils.getVerticalColor(localSettings, wikiVariables.vertical);
 	// the second argument is a whitelist of acceptable parameter names
