+++ server/facets/showCuratedContent.ts
@@ -5,9 +5,17 @@ import MainPage = require('../lib/MainPage');
 import MediaWiki = require('../lib/MediaWiki');
 import Utils = require('../lib/Utils');
 import localSettings = require('../../config/localSettings');
-import processCuratedContentData = require('./operations/processCuratedContentData');
+import prepareCuratedContentData = require('./operations/prepareCuratedContentData');
 import Promise = require('bluebird');
+import Caching = require('../lib/Caching');
+import Tracking = require('../lib/Tracking');
 
+var cachingTimes = {
+	enabled: true,
+	cachingPolicy: Caching.Policy.Public,
+	varnishTTL: Caching.Interval.standard,
+	browserTTL: Caching.Interval.disabled
+};
 
 export function category (request: Hapi.Request, reply: Hapi.Response): void {
 	fetchData(request, reply);
@@ -38,14 +46,14 @@ function fetchData(request: Hapi.Request, reply: Hapi.Response): void {
 	mainPage.getWikiVariablesAndDetails()
 		.then((data: CuratedContentPageData): void => {
 			Utils.redirectToCanonicalHostIfNeeded(localSettings, request, reply, data.wikiVariables);
-			processCuratedContentData(request, reply, data, allowCache);
+			outputResponse(request, reply, data, allowCache);
 		})
 		.catch(MainPage.MainPageDataRequestError, (error: any) => {
 			var data: CuratedContentPageData = error.data;
 			Logger.error('Error when fetching ads context and article details', data.mainPageData.exception);
-			processCuratedContentData(request, reply, data, false);
+			outputResponse(request, reply, data, false);
 		})
-		.catch(MediaWiki.WikiVariablesRequestError, (error: any) => {
+		.catch(MediaWiki.WikiVariablesRequestError, (error: any): void => {
 			Logger.error('Error when fetching wiki variables', error);
 			reply.redirect(localSettings.redirectUrlOnNoData);
 		})
@@ -56,4 +64,42 @@ function fetchData(request: Hapi.Request, reply: Hapi.Response): void {
 			Logger.fatal('Unhandled error, code issue', error);
 			reply.redirect(localSettings.redirectUrlOnNoData);
 		});
+}
+
+/**
+ * Handles article response from API
+ *
+ * @param {Hapi.Request} request
+ * @param {Hapi.Response} reply
+ * @param {ArticlePageData} data
+ * @param {boolean} allowCache
+ * @param {number} code
+ */
+function outputResponse(request: Hapi.Request,
+						reply: Hapi.Response,
+						data: CuratedContentPageData,
+						allowCache: boolean = true,
+						code: number = 200): void {
+	var response: Hapi.Response,
+		result: any;
+
+	if (!data.wikiVariables.dbName) {
+		//if we have nothing to show, redirect to our fallback wiki
+		reply.redirect(localSettings.redirectUrlOnNoData);
+	}
+
+	result = prepareCuratedContentData(request, data);
+
+	// @TODO XW-596 we shouldn't rely on side effects of this function
+	Tracking.handleResponse(result, request);
+
+	response = reply.view('application', result);
+	response.code(code);
+	response.type('text/html; charset=utf-8');
+
+	if (allowCache) {
+		return Caching.setResponseCaching(response, cachingTimes);
+	}
+
+	return Caching.disableCache(response);
 }
