+++ server/lib/MainPage.ts
@@ -65,7 +65,7 @@ export class MainPageRequestHelper {
 					wikiVariablesPromise.reason();
 
 				if (!isWikiVariablesPromiseFulfilled) {
-					return Promise.reject(new WikiVariablesRequestError(wikiVariables));
+					return Promise.reject(new MediaWiki.WikiVariablesRequestError(wikiVariables));
 				}
 
 				data = {
