+++ server/views/_layouts/ember-main.hbs
@@ -68,7 +68,7 @@
 				M.provide('article', {{{json article}}});
 			{{/if}}
 			{{#if mainPageData}}
-				M.prop('mainPageData', {{{json mainPageData}}}, true);
+				M.prop('mainPageData', {{{json mainPageData}}});
 			{{/if}}
 			{{#if exception}}
 				M.provide('exception', {{{json exception}}});
